+++ src/main/java/net/billstark001/worldmirror/ui/StatusClientScreen.java
@@ -4,7 +4,6 @@ import io.github.cottonmc.cotton.gui.client.CottonClientScreen;
 import net.billstark001.worldmirror.download.DownloadManager;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.minecraft.client.Minecraft;
 
 /**
  * Wraps {@link CottonClientScreen} around {@link StatusScreen} and
@@ -37,7 +36,7 @@ public class StatusClientScreen extends CottonClientScreen {
         if (currentExport != lastExportState || currentActive != lastActiveState) {
             lastExportState = currentExport;
             lastActiveState = currentActive;
-            Minecraft.getInstance().setScreen(new StatusClientScreen());
+            ClientScreens.set(new StatusClientScreen());
         }
     }
 }